test base refactoring
