+++ server/tests/__init__.py
@@ -1,0 +1,2 @@
+import os
+os.environ['APP_SETTINGS'] = 'server.config.TestingConfig'
